+++ innovation_sweet_spots/analysis/prototyping/21_tomotopy.py
@@ -141,26 +141,6 @@ for doc in full_corpus:
 original_indexes = list(range(len(corpus_gtr) + len(corpus_cb)))
 original_indexes = [i for i in original_indexes if i not in empty_docs]
 
-# %%
-
-# %%
-len(original_indexes)
-
-# %%
-import innovation_sweet_spots.analysis.analysis_utils as iss
-import innovation_sweet_spots.analysis.topic_analysis as iss_topics
-
-doc_df = pd.DataFrame(data={"doc_id": full_corpus_ids})
-doc_df = iss_topics.get_doc_details(doc_df)
-
-# %%
-doc_df["top2vec_index"] = original_indexes
-
-# %%
-doc_df[["doc_id", "title", "source", "top2vec_index"]].to_csv(
-    PROJECT_DIR / "outputs/models/topic_models/topic_model_docs.csv", index=False
-)
-
 # %% [markdown]
 # # Global topic model
 
@@ -194,6 +174,8 @@ mdl = tp.LDAModel.load(str(PROJECT_DIR / "outputs/models/full_corpus_lda_model.b
 # ## Inspect tomotopy results
 
 # %%
+import innovation_sweet_spots.analysis.analysis_utils as iss
+import innovation_sweet_spots.analysis.topic_analysis as iss_topics
 
 # %%
 import importlib
@@ -352,10 +334,10 @@ df = doc_df_gtr[doc_df_gtr.text == "Energy Storage"]
 # doc_df_[doc_df_.doc_id.isin(df.project_id.to_list())].prob.mean()
 
 # %%
-# prob_histogram(df_probs)
+prob_histogram(df_probs)
 
 # %%
-# fig = cumulative_plot(df_probs)
+fig = cumulative_plot(df_probs)
 # plt.xscale('log')
 
 # %%
@@ -549,10 +531,6 @@ len(full_corpus)
 
 # %%
 import innovation_sweet_spots.getters.crunchbase as crunchbase
-
-# %%
-# Doc dataframe
-narrow_set_df = doc_df_.copy()
 
 # %%
 doc_df_gtr = doc_df[doc_df.source == "gtr"]
@@ -779,9 +757,6 @@ for cat in cats:
     ].topic.to_list()
 
 # %%
-categories_to_narrow_topics
-
-# %%
 global_topics = [26, 52, 68, 71, 78, 83, 128, 139, 140, 142]
 local_topics = [
     40,
@@ -829,26 +804,21 @@ from tqdm.notebook import tqdm
 #     infer_doc_topic_dists.append(infer_doc_topic_dist)
 
 # %%
-# infer_doc_topic_dists = []
-# for doc in tqdm(full_corpus, total=len(full_corpus)):
-#     doc_instance = mdl_new.make_doc(doc)
-#     infer_doc_topic_dist, _ = mdl_new.infer(doc_instance)
-#     infer_doc_topic_dists.append(infer_doc_topic_dist)
-# infer_doc_topic_dists = np.array(infer_doc_topic_dists)
+infer_doc_topic_dists = []
+for doc in tqdm(full_corpus, total=len(full_corpus)):
+    doc_instance = mdl_new.make_doc(doc)
+    infer_doc_topic_dist, _ = mdl_new.infer(doc_instance)
+    infer_doc_topic_dists.append(infer_doc_topic_dist)
+infer_doc_topic_dists = np.array(infer_doc_topic_dists)
 
 # %%
 # np.save(PROJECT_DIR / 'outputs/data/results_august/narrow_topics_full_corpus.npy', np.array(infer_doc_topic_dists))
 
 # %%
-infer_doc_topic_dists = np.load(
-    PROJECT_DIR / "outputs/data/results_august/narrow_topics_full_corpus.npy"
-)
-
-# %%
 infer_doc_topic_dists.shape
 
 # %%
-# global_doc_topic_probs.shape
+global_doc_topic_probs.shape
 
 # %%
 # check_docs_id_ = check_docs_id
@@ -958,34 +928,18 @@ for df in selected_docs_dfs:
     selected_docs_df_all = selected_docs_df_all.append(df, ignore_index=True)
 
 # %%
-# selected_docs_df_all.to_csv(PROJECT_DIR / 'outputs/data/results_august/reference_categories_to_check.csv', index=False)
-
-# %%
-selected_docs_df_all = pd.read_csv(
-    PROJECT_DIR / "outputs/data/results_august/reference_categories_to_check.csv"
+selected_docs_df_all.to_csv(
+    PROJECT_DIR / "outputs/data/results_august/reference_categories_to_check.csv",
+    index=False,
 )
-
-# %%
-selected_docs_df_all[
-    selected_docs_df_all.doc_id == "34FFD228-2F72-461E-AD54-A92D2D826225"
-]
-
-# %%
-print(selected_docs_df_all.iloc[1491].description)
 
 # %%
 np.sum([len(df) for df in selected_docs_dfs])
 
 # %%
-cats
-
-# %%
-i = 5
+i = 2
 print(cats[i])
 selected_docs_dfs[i].groupby("selection").count()
-
-# %%
-[len(df) for df in selected_docs_dfs]
 
 # %%
 selected_docs_dfs[i]
@@ -1208,10 +1162,6 @@ narrow_set_df.loc[ids].probs.min()
 # # Keyword approach
 
 # %%
-# Doc dataframe
-narrow_set_df = doc_df_.copy()
-
-# %%
 make_pyLDAvis(
     mdl_new,
     fpath=PROJECT_DIR / "outputs/data/results_august/ldavis_tomotopy_narrow_set.html",
@@ -1223,31 +1173,12 @@ new_corpus_document_texts = [
     " ".join([" ".join(s.split("_")) for s in doc]) for doc in new_corpus
 ]
 # Where are new document ids?
-# narrow_set_df.head(1)
+narrow_set_df.head(1)
 
 # %%
 full_corpus_document_texts = [
-    " " + " ".join([" ".join(s.split("_")) + " " for s in doc]) for doc in full_corpus
+    " ".join([" ".join(s.split("_")) for s in doc]) for doc in full_corpus
 ]
-
-# %%
-# full_corpus_document_texts_2 = [' ' + ' '.join([' '.join(s.split('_')) + ' ' for s in doc]) for doc in full_corpus]
-
-# %%
-full_corpus_document_texts_2 = []
-for tokens_list in full_corpus:
-    s_list = [s for s_list in [tok.split("_") for tok in tokens_list] for s in s_list]
-    full_corpus_document_texts_2.append(" " + " ".join(s_list) + " ")
-
-# %%
-full_corpus_document_texts[0]
-
-# %%
-full_corpus_document_texts_2[0]
-
-# %%
-global_doc_topic_dists = get_doc_topic_dists(mdl)
-local_doc_topic_dists = get_doc_topic_dists(mdl_new)
 
 
 # %%
@@ -1300,19 +1231,6 @@ def get_docs_with_keyphrases(keyphrases):
     return doc_df.iloc[x]
 
 
-def get_docs_with_keyphrases_2(keyphrases):
-    x = np.array([False] * len(full_corpus_document_texts_2))
-    for terms in keyphrases:
-        print(terms)
-        x = x | find_docs_with_all_terms(
-            terms,
-            corpus_texts=full_corpus_document_texts_2,
-            corpus_df=doc_df,
-            return_dataframe=False,
-        )
-    return doc_df.iloc[x]
-
-
 def add_info(df, GLOBAL_TOPIC, LOCAL_TOPIC):
     df = df.merge(clustering, how="left")
 
@@ -1347,8 +1265,6 @@ def print_categories_keywords(d):
 
 
 # %%
-
-# %%
 categories_keyphrases_heat = {
     # Heating subcategories
     "Heat pumps": [["heat pump"]],
@@ -1380,19 +1296,14 @@ categories_keyphrases_build = {
     "Building insulation": [
         ["insulat", "build"],
         ["insulat", "hous"],
-        ["insulat", "home"],
         ["insulat", "retrofit"],
         ["cladding", "hous"],
         ["cladding", "build"],
         ["glazing", "window"],
         ["glazed", "window"],
     ],
-    "Building insulation (extra)": [
-        ["retrofit", "build"],
-        ["retrofit", "hous"],
-        ["retrofit", "home"],
-    ],
     "Radiators": [["radiator"]],
+    #     'BIM': [[' bim '], ['building information management']],
 }
 
 categories_keyphrases_energy = {
@@ -1403,15 +1314,13 @@ categories_keyphrases_energy = {
         ["thermostat"],
         ["smart meter"],
     ],
-    "Energy management (extra)": [["load shift", "heat"]],
+    "Energy management (extra)": [["energy management", "home"]],
     "Energy management": [
         ["smart home", "heat"],
         ["demand response", "heat"],
-        ["load shift", "heat"],
         ["energy management", "build"],
         ["energy management", "domestic"],
         ["energy management", "hous"],
-        ["energy management", "home"],
         ["thermostat"],
         ["smart meter"],
     ],
@@ -1420,9 +1329,6 @@ categories_keyphrases_energy = {
 #     'Heat transfer': [['heat transfer'], ['heat exchange']],
 #     'Cooling and refrigeration': [['cooling'], ['refriger']],
 # 'Other heating' category
-
-# %% [markdown]
-# #### Extra checks
 
 # %%
 df = get_docs_with_keyphrases(categories_keyphrases_heat["Hydrogen heating"])
@@ -1438,22 +1344,6 @@ df = add_info(df, GLOBAL_TOPIC=68, LOCAL_TOPIC=19).reset_index(drop=True)
 df.to_csv(
     PROJECT_DIR
     / "outputs/data/aux/ISS_technologies_to_review_August_10 - Hydrogen heating.csv"
-)
-
-# %%
-df_check = pd.read_excel(
-    PROJECT_DIR
-    / "outputs/data/results_august/aux/ISS_technologies_to_review_August_10.xlsx",
-    sheet_name="Building insulation",
-)
-df = get_docs_with_keyphrases(
-    categories_keyphrases_build["Building insulation (extra)"]
-)
-df = df[df.doc_id.isin(df_check.doc_id.to_list()) == False]
-df = add_info(df, GLOBAL_TOPIC=140, LOCAL_TOPIC=149).reset_index(drop=True)
-df.to_csv(
-    PROJECT_DIR
-    / "outputs/data/aux/ISS_technologies_to_review_August_10 - Building insulation (extra).csv"
 )
 
 # %%
@@ -1473,13 +1363,15 @@ df.to_csv(
 )
 
 # %%
-df_check = pd.read_excel(
-    PROJECT_DIR
-    / "outputs/data/results_august/aux/ISS_technologies_to_review_August_10.xlsx",
-    sheet_name="Energy management",
-)
 df = get_docs_with_keyphrases(categories_keyphrases_energy["Energy management (extra)"])
-df = df[df.doc_id.isin(df_check.doc_id.to_list()) == False]
+df = df[
+    df.doc_id.isin(
+        get_docs_with_keyphrases(
+            categories_keyphrases_energy["Energy management"]
+        ).doc_id.to_list()
+    )
+    == False
+]
 df = add_info(df, GLOBAL_TOPIC=140, LOCAL_TOPIC=149).reset_index(drop=True)
 df.to_csv(
     PROJECT_DIR
@@ -1487,48 +1379,7 @@ df.to_csv(
 )
 
 # %%
-df_check = pd.read_excel(
-    PROJECT_DIR
-    / "outputs/data/results_august/aux/ISS_technologies_to_review_August_10.xlsx",
-    sheet_name="Building insulation",
-)
-df = get_docs_with_keyphrases(
-    categories_keyphrases_build["Building insulation (extra)"]
-)
-df = df[df.doc_id.isin(df_check.doc_id.to_list()) == False]
-df = add_info(df, GLOBAL_TOPIC=140, LOCAL_TOPIC=149).reset_index(drop=True)
-df.to_csv(
-    PROJECT_DIR
-    / "outputs/data/aux/ISS_technologies_to_review_August_10 - Building insulation (extra).csv"
-)
-
-# %%
-GTR_DOCS_ALL = pd.read_csv(
-    PROJECT_DIR / "outputs/data/results_august/checked_heating_tech_GTR.csv"
-)
-
-# %%
-len(df)
-
-# %%
-df = df[
-    df.doc_id.isin(
-        GTR_DOCS_ALL[
-            GTR_DOCS_ALL.tech_category == "Building insulation"
-        ].doc_id.to_list()
-    )
-    == False
-]
-df.to_csv(
-    PROJECT_DIR
-    / "outputs/data/aux/ISS_technologies_to_review_August_10 - Building insulation (extra).csv"
-)
-
-# %%
-# df = get_docs_with_keyphrases(categories_keyphrases_energy['Energy management'])
-# df = df[df.doc_id.isin(get_docs_with_keyphrases(categories_keyphrases_energy['Energy management']).doc_id.to_list())==False]
-# df = add_info(df, GLOBAL_TOPIC=140, LOCAL_TOPIC = 149).reset_index(drop=True)
-# df.to_csv(PROJECT_DIR / 'outputs/data/aux/ISS_technologies_to_review_August_10 - Energy management (extra).csv')
+df
 
 # %%
 # df = pd.read_csv(OUTPUTS_DIR.parent / 'aux/ISS_technologies_to_review_August_10 - Electric boilers.csv')
@@ -1569,11 +1420,12 @@ print_categories_keywords(categories_keyphrases_build)
 print_categories_keywords(categories_keyphrases_energy)
 
 # %%
+global_doc_topic_dists = get_doc_topic_dists(mdl)
+local_doc_topic_dists = get_doc_topic_dists(mdl_new)
+
+# %%
 GLOBAL_TOPIC = 68
 LOCAL_TOPIC = 19
-
-# %% [markdown]
-# #### Initial checks
 
 # %%
 # df = get_docs_with_keyphrases(categories_keyphrases['Biomass boilers'])
@@ -1613,1415 +1465,7 @@ with pd.ExcelWriter(
     for key in dfs:
         dfs[key].to_excel(writer, sheet_name=key)
 
-# %% [markdown]
-# ### Check spacing and keyword search differences
-
-# %%
-# df = get_docs_with_keyphrases(categories_keyphrases['Biomass boilers'])
-dfs_1 = {
-    key: add_info(
-        get_docs_with_keyphrases(categories_keyphrases_heat[key]),
-        GLOBAL_TOPIC=68,
-        LOCAL_TOPIC=19,
-    ).reset_index(drop=True)
-    for key in categories_keyphrases_heat
-}
-dfs_2 = {
-    key: add_info(
-        get_docs_with_keyphrases_2(categories_keyphrases_heat[key]),
-        GLOBAL_TOPIC=68,
-        LOCAL_TOPIC=19,
-    ).reset_index(drop=True)
-    for key in categories_keyphrases_heat
-}
-
-for key in categories_keyphrases_build:
-    dfs_1[key] = add_info(
-        get_docs_with_keyphrases(categories_keyphrases_build[key]),
-        GLOBAL_TOPIC=140,
-        LOCAL_TOPIC=149,
-    ).reset_index(drop=True)
-    dfs_2[key] = add_info(
-        get_docs_with_keyphrases_2(categories_keyphrases_build[key]),
-        GLOBAL_TOPIC=140,
-        LOCAL_TOPIC=149,
-    ).reset_index(drop=True)
-
-for key in categories_keyphrases_energy:
-    dfs_1[key] = add_info(
-        get_docs_with_keyphrases(categories_keyphrases_energy[key]),
-        GLOBAL_TOPIC=68,
-        LOCAL_TOPIC=138,
-    ).reset_index(drop=True)
-    dfs_2[key] = add_info(
-        get_docs_with_keyphrases_2(categories_keyphrases_build[key]),
-        GLOBAL_TOPIC=140,
-        LOCAL_TOPIC=149,
-    ).reset_index(drop=True)
-
-# %%
-print([len(dfs_1[df]) for df in dfs_1])
-print([len(dfs_2[df]) for df in dfs_2])
-
-# %%
-dfs_2_removed = {}
-for key in dfs_1:
-    df = dfs_1[key]
-    df_new = dfs_2[key][dfs_2[key].doc_id.isin(df.doc_id.to_list()) == False]
-    if len(df_new) != 0:
-        dfs_2_removed[key] = df_new
-
-# %%
-for key in dfs_2_removed:
-    print(key, len(dfs_2_removed[key]))
-
-# %%
-with pd.ExcelWriter(
-    PROJECT_DIR
-    / "outputs/data/results_august/ISS_technologies_to_review_September_1.xlsx"
-) as writer:
-    for key in dfs_2_removed:
-        dfs_2_removed[key].to_excel(writer, sheet_name=key)
-
 # %%
 # dfs['Heat pumps']
-
-# %% [markdown]
-# ### Keywords and reference categories
-
-# %%
-cats
-
-# %%
-df_labels = pd.read_csv(
-    "/Users/karliskanders/Downloads/project-7-at-2021-08-20-11-17-039f3681.csv"
-)
-
-# %%
-# category = 'Bioenergy'
-# category= 'Nuclear'
-category = "EV"
-# category = 'Wind & Offshore'
-# category = 'Carbon Capture & Storage'
-# category = 'Hydrogen & Fuel Cells'
-# df = selected_docs_df_all[selected_docs_df_all.category=='Hydrogen & Fuel Cells']
-df = selected_docs_df_all[selected_docs_df_all.category == category]
-corpus_docs_ids = doc_df[doc_df.doc_id.isin(df.doc_id.to_list())].index.to_list()
-
-# %%
-df_labels[df_labels.category == category].groupby(["selection", "sentiment"]).count()
-
-# %%
-# full_corpus_document_texts[0]
-
-# %%
-# dff = get_docs_with_keyphrases([['solar'], [' pv '], ['perovskite'], ['photovoltaic']])
-# dff = get_docs_with_keyphrases([['battery'], ['batteries']])
-# dff = get_docs_with_keyphrases([['biofuel'], ['syngas'], ['bioenergy'], ['biogas'], ['biochar'],
-#                                 ['biodiesel'], ['bioethanol'], ['biobutanol'], ['biomass', 'energy'], ['biomass', 'power'],
-#                                 ['anaerobic digestion'], ['torrefaction'], ['pyrolisis'], ['ethanol'], ['butanol']
-#                                ])
-# dff = get_docs_with_keyphrases([['hydrogen'], ['fuel cell'], ['sofc']])
-dff = get_docs_with_keyphrases([["ccs"], ["carbon storage"], ["carbon capture"]])
-
-# %%
-df_labels_cat = df_labels[df_labels.category == category]
-df_labels_cat["has_keyword"] = 0
-df_labels_cat.loc[df_labels_cat.doc_id.isin(dff.doc_id.to_list()), "has_keyword"] = 1
-
-# %%
-df_check = df_labels_cat.groupby(["source", "sentiment", "has_keyword"]).count()
-df_check
-
-# %%
-# cats_to_numbers = dict(zip(cats, range(len(cats))))
-
-# %%
-# selected_docs = [full_corpus[doc_df[doc_df.doc_id==doc_id].index[0]] for doc_id in df.doc_id.to_list()]
-
-# %%
-# full_corpus_dict = dict(zip(full_corpus_ids, full_corpus))
-
-# %%
-# x=[]
-# for doc_id in df.doc_id.to_list():
-#     if doc_id in full_corpus_dict:
-#         x.append(full_corpus_dict[doc_id])
-#     else:
-#         x.append([''])
-#         print(doc_id)
-
-# %%
-# df = selected_docs_df_all.copy()
-# df['category_n'] = df.category.apply(lambda x: cats_to_numbers[x])
-# iss_topics.topic_keywords(x, df.category_n.to_list(), topic_words='bigrams')
-
-# %%
-# print(df_labels_cat[(df_labels_cat.sentiment=='Discard')&(df_labels_cat.has_keyword==1)].sample().text.iloc[0])
-
-
-# %%
-# dff = get_docs_with_keyphrases([[' hydrogen '], ['fuel cell'], ['sofc']])
-# dff = get_docs_with_keyphrases([[' hydrogen '], ['fuel cell'], ['sofc']])
-# dff = get_docs_with_keyphrases([['ccs'],
-#                                 ['carbon storage'],
-#                                 ['carbon capture'],
-#                                 ['co2 capture'],
-#                                 ['post combustion capture'],
-#                                 ['carbon abatement'],
-#                                 ['flue gas']
-#                                ])
-# dff = get_docs_with_keyphrases([['wind', 'energy'],
-#                                 ['wind', 'renewable'],
-#                                 ['wind turbine'],
-#                                 ['wind farm'],
-#                                 ['wind generator'],
-#                                 ['offshore wind'],
-#                                 ['onshore wind'],
-#                                 ['wind power'],
-#                                 ['wave', 'renewable'],
-#                                 ['wave', 'energy'],
-#                                 ['wave power'],
-#                                 ['tidal', 'energy'],
-#                                 ['tidal', 'turbine'],
-#                                 ['marine', 'renewable'],
-#                                 ['marine' , 'energy'],
-#                                 ['offshore energy'],
-#                                ])
-# dff = get_docs_with_keyphrases([['nuclear'],
-#                                 ['radioactive', 'waste'],
-#                                 ['fusion', 'power'],
-#                                 ['fusion', 'reactor'],
-#                                 ['tokamak']
-#                                ])
-# dff = get_docs_with_keyphrases([['electric vehicle'],
-# #                                 [' ev '],
-# #                                 ['electric', 'vehicle'],
-#                                ])
-
-# %%
-len(df)
-
-# %%
-df_cat = df[df.category == category]
-df_cat["has_keyword"] = 0
-df_cat.loc[df_cat.doc_id.isin(dff.doc_id.to_list()), "has_keyword"] = 1
-
-# %%
-df_cat.has_keyword.sum()
-
-# %%
-df_check = df_cat.groupby(
-    list(reversed(["selection", "source", "has_keyword"]))
-).count()
-df_check
-
-# %%
-# full_corpus_dict['372CD65D-7E86-4E64-9630-03843144D797']
-i = 0
-
-# %%
-i += 1
-print(
-    df_cat[(df_cat.selection == "top2vec") & (df_cat.has_keyword != 1)]
-    .sort_values("topic_prob", ascending=False)
-    .iloc[i]
-    .text
-)
-
-
-# %%
-len(dff[dff.doc_id.isin(df_cat.doc_id.to_list()) == False])
-
-# %%
-dff[dff.doc_id.isin(df_cat.doc_id.to_list()) == False].sample()[
-    ["title", "description"]
-].iloc[0]
-
-# %%
-# df_cat[df_cat.title=='Lithium Innovations for Future Electric vehicles (LIFE)']
-
-# %%
-words, word_scores = top2vec_model.similar_words(
-    keywords=["sofc"], keywords_neg=[], num_words=25
-)
-for word, score in zip(words, word_scores):
-    print(f"{word} {score}")
-
-# %% [markdown]
-# # Keyword+
-
-# %%
-# df_manual_labels = pd.read_csv('/Users/karliskanders/Downloads/project-7-at-2021-08-20-11-17-039f3681.csv')
-df_manual_labels = pd.read_csv(
-    PROJECT_DIR
-    / "outputs/data/results_august/aux/project-7-at-2021-08-20-11-17-039f3681.csv"
-)
-
-# %%
-ref_keywords = {
-    "Carbon Capture & Storage": [
-        [" ccs "],
-        ["carbon storage"],
-        ["carbon capture"],
-        ["co2 capture"],
-        ["post combustion capture"],
-        ["carbon abatement"],
-        ["flue gas"],
-    ],
-    "Bioenergy": [
-        ["bioenergy"],
-        ["biofuel"],
-        ["syngas"],
-        ["biogas"],
-        ["biochar"],
-        ["biodiesel"],
-        ["biomass", "energy"],
-        ["biomass", "power"],
-        ["anaerobic digestion"],
-        ["torrefaction"],
-        ["pyrolisis"],
-        ["ethanol", "fuel"],
-        ["butanol", "fuel"],
-    ],
-    "Nuclear": [
-        ["nuclear"],
-        ["radioactive", "waste"],
-        ["fusion", "power"],
-        ["fusion", "reactor"],
-        ["tokamak"],
-    ],
-    "Hydrogen & Fuel Cells": [[" hydrogen "], ["fuel cell"], ["sofc"]],
-    "Wind & Offshore": [
-        ["wind", "energy"],
-        ["wind", "renewable"],
-        ["wind turbine"],
-        ["wind farm"],
-        ["wind generator"],
-        ["offshore wind"],
-        ["onshore wind"],
-        ["wind power"],
-        ["wave", "renewable"],
-        ["wave", "energy"],
-        ["wave power"],
-        ["tidal", "energy"],
-        ["tidal", "turbine"],
-        ["marine", "renewable"],
-        ["marine", "energy"],
-        ["offshore energy"],
-    ],
-    "EV": [
-        ["electric vehicle"],
-        [" ev "],
-        ["electric", "vehicle"],
-    ],
-    "Heating & Buildings": [
-        ["infrared", "heating"],
-        ["boiler"],
-        ["heat"],
-    ],
-}
-
-# %% [markdown]
-# - Check selected documents that don't have the keywords
-# - Check unselected documents that have (perhaps a narrower version) of the keywords
-
-# %%
-# category = 'Wind & Offshore'
-# dff = get_docs_with_keyphrases(ref_keywords[category])
-# df_cat = selected_docs_df_all[selected_docs_df_all.category==category]
-# df_cat['has_keyword'] = 0
-# df_cat.loc[df_cat.doc_id.isin(dff.doc_id.to_list()), 'has_keyword'] = 1
-
-# %%
-# len(dff), len(df_cat), df_cat.has_keyword.sum()
-
-# %%
-# i+=1
-# # print(df_cat[(df_cat.selection=='top2vec') & (df_cat.has_keyword==1)].sort_values('topic_prob', ascending=False).iloc[i].text)
-
-
-# %% [markdown]
-# # Unselected documents that have the keywords
-
-# %%
-# df_unselected_key = dff[dff.doc_id.isin(df_cat.doc_id.to_list())==False]
-
-# %%
-narrow_ref_keywords = {
-    "Batteries": [["battery"], ["batteries"]],
-    "Solar": [
-        ["solar energy"],
-        ["solar", "renewable"],
-        ["solar cell"],
-        ["solar panel"],
-        ["photovoltaic"],
-        ["perovskite"],
-        ["pv cell"],
-    ],
-    "Carbon Capture & Storage": [
-        [" ccs "],
-        ["carbon storage"],
-        ["carbon capture"],
-        ["co2 capture"],
-        ["carbon abatement"],
-    ],
-    "Bioenergy": [
-        ["biofuel"],
-        ["ethanol", "fuel"],
-        ["butanol", "fuel"],
-        ["syngas"],
-        ["biogas"],
-        ["biochar"],
-        ["biodiesel"],
-        ["torrefaction", "biomass"],
-        ["pyrolisis", "biomass"],
-    ],
-    "Nuclear": [
-        ["nuclear", "energy"],
-        ["radioactive", "waste"],
-        ["nuclear", "fission"],
-        ["fission", "reactor"],
-        ["nuclear", "fusion"],
-        ["fusion", "reactor"],
-        ["tokamak"],
-    ],
-    "Hydrogen & Fuel Cells": [[" hydrogen "], ["fuel cell"], ["sofc"]],
-    "Wind & Offshore": [
-        ["wind energy"],
-        ["wind", "renewable"],
-        ["wind turbine"],
-        ["wind farm"],
-        ["wind generator"],
-        ["wind power"],
-        ["wave", "renewable"],
-        ["tidal energy"],
-        ["tidal turbine"],
-        ["offshore energy"],
-        ["offshore wind"],
-        ["onshore wind"],
-    ],
-    "EV": [
-        ["electric vehicle"],
-        [" ev "],
-        ["electric", "vehicle"],
-    ],
-    "Heating (all other)": [
-        ["infrared", "heating"],
-        ["heat", "build"],
-        ["heat", "home"],
-        ["heat", "hous"],
-        ["heat", "domestic"],
-        ["heat", "renewable"],
-        ["heating"],
-    ],
-    "Demand management": [
-        [" vpp "],
-        ["virtual power plant"],
-        ["smart building"],
-        ["smart home"],
-        ["smart meter"],
-        ["smart grid"],
-        ["demand response"],
-    ],
-}
-
-# %%
-narrow_topic_probs = np.load(
-    PROJECT_DIR / "outputs/data/results_august/narrow_topics_full_corpus.npy"
-)
-
-# %%
-SMALL_THRESH = 0.01
-
-# %%
-categories_to_narrow_topics
-
-# %% [markdown]
-# ## Solar: false negative check
-
-# %%
-# NB: This is for selecting the new docs to check
-manual_cb_review = pd.read_csv(
-    PROJECT_DIR
-    / "outputs/data/results_august/aux/project-11-at-2021-08-24-17-28-6e9a63e0.csv"
-)
-
-# %%
-DF_NEW = {}
-
-# %%
-category = "Solar"
-categories_to_narrow_topics[category]
-
-# %%
-category_reviewed = df_manual_labels[df_manual_labels.category == category]
-category_reviewed_hits = category_reviewed[category_reviewed.sentiment == "Keep"]
-category_reviewed_misses = category_reviewed[category_reviewed.sentiment == "Discard"]
-
-# %%
-# Narrow topic probability
-doc_df_ = doc_df[["doc_id"]].copy()
-doc_df_["topic_probs"] = narrow_topic_probs[:, 2]
-# All documents with the keywords
-df_with_keywords_old = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords = get_docs_with_keyphrases_2(narrow_ref_keywords[category])
-print(len(df_with_keywords_old), len(df_with_keywords))
-# Above-threshold, selected documents that have the keyword
-df_selected = selected_docs_df_all[selected_docs_df_all.category == category]
-df_selected["has_keyword"] = 0
-df_selected.loc[
-    df_selected.doc_id.isin(df_with_keywords.doc_id.to_list()), "has_keyword"
-] = 1
-# Documents with keywords that were not selected
-df_unselected_with_keywords = df_with_keywords[
-    df_with_keywords.doc_id.isin(df_selected.doc_id.to_list()) == False
-]
-df_unselected_with_keywords = df_unselected_with_keywords.merge(doc_df_, how="left")
-
-# %%
-len(df_selected), df_selected.has_keyword.sum(), len(df_unselected_with_keywords)
-
-# %%
-# category_reviewed_hits['doc_id']
-pd.set_option("max_colwidth", 140)
-
-# %%
-df_ = category_reviewed_hits[["doc_id", "title", "description", "source"]]
-df_ = df_.merge(doc_df_)
-df_["manual_ok"] = 1
-df_new = df_unselected_with_keywords
-df_new["manual_ok"] = 0
-df_ = df_.append(df_new, ignore_index=True)
-df_ = df_.merge(clustering[["doc_id", "cluster_keywords"]])
-df_["category"] = category
-
-data_for_labelling = []
-for i, row in df_.iterrows():
-    s = f"CATEGORY: {category}\n\n"
-    s += f"doc_id: {row.doc_id}\n\nprobability: {np.round(row.topic_probs,4)}\n\n{row.title}\n------\n{row.description}\n------"
-    s += f"\n\ntop2vec keywords: {row.cluster_keywords}"
-    data_for_labelling.append(s)
-df_["text"] = data_for_labelling
-
-# %%
-df_new_cb = df_[(df_.source == "cb") & (df_.manual_ok == 0)]
-df_new_cb.to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Aug23_cb_{category}.csv",
-    index=False,
-)
-
-df_new_cb[df_new_cb.doc_id.isin(manual_cb_review.doc_id.to_list()) == False].to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Sep1_cb_{category}.csv",
-    index=False,
-)
-
-DF_NEW[category] = df_[
-    (df_.source == "gtr") | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-
-# %%
-print(len(DF_NEW[category]))
-DF_NEW[category] = df_[
-    (
-        (df_.source == "gtr")
-        & ((df_.manual_ok == 1) | df_.doc_id.isin(df_with_keywords.doc_id.to_list()))
-    )
-    | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-len(DF_NEW[category])
-
-# %%
-len(df_), len(df_new), len(df_new_cb)
-
-# %% [markdown]
-# ## Wind: false positive and negative check
-
-# %%
-category = "Wind & Offshore"
-categories_to_narrow_topics[category]
-
-# %%
-# Narrow topic probability
-doc_df_ = doc_df[["doc_id"]].copy()
-doc_df_["topic_probs"] = narrow_topic_probs[:, 67]
-# All documents with the keywords
-# df_with_keywords = get_docs_with_keyphrases(narrow_ref_keywords[category])
-
-df_with_keywords_old = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords = get_docs_with_keyphrases_2(narrow_ref_keywords[category])
-print(len(df_with_keywords_old), len(df_with_keywords))
-
-# Above-threshold, selected documents that have the keyword
-df_selected = selected_docs_df_all[selected_docs_df_all.category == category]
-df_selected["has_keyword"] = 0
-df_selected.loc[
-    df_selected.doc_id.isin(df_with_keywords.doc_id.to_list()), "has_keyword"
-] = 1
-# Documents with keywords that were not selected
-df_unselected_with_keywords = df_with_keywords[
-    df_with_keywords.doc_id.isin(df_selected.doc_id.to_list()) == False
-]
-df_unselected_with_keywords = df_unselected_with_keywords.merge(doc_df_, how="left")
-
-# %%
-df_ = df_selected[["doc_id", "title", "description", "source"]]
-df_ = df_.merge(doc_df_)
-df_["manual_ok"] = 0
-df_new = df_unselected_with_keywords
-df_new["manual_ok"] = 0
-df_ = df_.append(df_new, ignore_index=True)
-df_ = df_.merge(clustering[["doc_id", "cluster_keywords"]])
-df_["category"] = category
-
-data_for_labelling = []
-for i, row in df_.iterrows():
-    s = f"CATEGORY: {category}\n\n"
-    s += f"doc_id: {row.doc_id}\n\nprobability: {np.round(row.topic_probs,4)}\n\n{row.title}\n------\n{row.description}\n------"
-    s += f"\n\ntop2vec keywords: {row.cluster_keywords}"
-    data_for_labelling.append(s)
-df_["text"] = data_for_labelling
-
-# %%
-df_new_cb = df_[(df_.source == "cb") & (df_.manual_ok == 0)]
-df_new_cb.to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Aug23_cb_{category}.csv",
-    index=False,
-)
-
-df_new_cb[df_new_cb.doc_id.isin(manual_cb_review.doc_id.to_list()) == False].to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Sep1_cb_{category}.csv",
-    index=False,
-)
-
-DF_NEW[category] = df_[
-    (df_.source == "gtr") | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-
-# %%
-print(len(DF_NEW[category]))
-DF_NEW[category] = df_[
-    (
-        (df_.source == "gtr")
-        & ((df_.manual_ok == 1) | df_.doc_id.isin(df_with_keywords.doc_id.to_list()))
-    )
-    | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-len(DF_NEW[category])
-
-# %%
-len(DF_NEW[category])
-
-# %%
-len(df_new_cb)
-
-# %%
-len(df_selected), df_selected.has_keyword.sum(), len(df_unselected_with_keywords)
-
-# %%
-len(df_unselected_with_keywords[df_unselected_with_keywords.topic_probs > 0.05])
-
-# %% [markdown]
-# ## Batteries: false negative check
-
-# %%
-category = "Batteries"
-categories_to_narrow_topics[category]
-
-# %%
-category_reviewed = df_manual_labels[df_manual_labels.category == category]
-category_reviewed_hits = category_reviewed[category_reviewed.sentiment == "Keep"]
-category_reviewed_misses = category_reviewed[category_reviewed.sentiment == "Discard"]
-
-# %%
-# Narrow topic probability
-doc_df_ = doc_df[["doc_id"]].copy()
-doc_df_["topic_probs"] = narrow_topic_probs[:, 40]
-# All documents with the keywords
-# df_with_keywords = get_docs_with_keyphrases(narrow_ref_keywords[category])
-
-df_with_keywords_old = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords = get_docs_with_keyphrases_2(narrow_ref_keywords[category])
-print(len(df_with_keywords_old), len(df_with_keywords))
-
-# Above-threshold, selected documents that have the keyword
-df_selected = selected_docs_df_all[selected_docs_df_all.category == category]
-df_selected["has_keyword"] = 0
-df_selected.loc[
-    df_selected.doc_id.isin(df_with_keywords.doc_id.to_list()), "has_keyword"
-] = 1
-# Documents with keywords that were not selected
-df_unselected_with_keywords = df_with_keywords[
-    df_with_keywords.doc_id.isin(df_selected.doc_id.to_list()) == False
-]
-df_unselected_with_keywords = df_unselected_with_keywords.merge(doc_df_, how="left")
-
-# %%
-df_ = category_reviewed_hits[["doc_id", "title", "description", "source"]]
-df_ = df_.merge(doc_df_)
-df_["manual_ok"] = 1
-df_new = df_unselected_with_keywords
-df_new["manual_ok"] = 0
-df_ = df_.append(df_new, ignore_index=True)
-df_ = df_.merge(clustering[["doc_id", "cluster_keywords"]])
-df_["category"] = category
-
-data_for_labelling = []
-for i, row in df_.iterrows():
-    s = f"CATEGORY: {category}\n\n"
-    s += f"doc_id: {row.doc_id}\n\nprobability: {np.round(row.topic_probs,4)}\n\n{row.title}\n------\n{row.description}\n------"
-    s += f"\n\ntop2vec keywords: {row.cluster_keywords}"
-    data_for_labelling.append(s)
-df_["text"] = data_for_labelling
-
-# %%
-df_new_cb = df_[(df_.source == "cb") & (df_.manual_ok == 0)]
-df_new_cb.to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Aug23_cb_{category}.csv",
-    index=False,
-)
-DF_NEW[category] = df_[
-    (df_.source == "gtr") | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-
-# %%
-print(len(DF_NEW[category]))
-DF_NEW[category] = df_[
-    (
-        (df_.source == "gtr")
-        & ((df_.manual_ok == 1) | df_.doc_id.isin(df_with_keywords.doc_id.to_list()))
-    )
-    | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-len(DF_NEW[category])
-
-# %%
-len(df_selected), df_selected.has_keyword.sum(), len(df_unselected_with_keywords)
-
-# %%
-len(df_unselected_with_keywords[df_unselected_with_keywords.topic_probs > 0.05])
-
-# %%
-# df_unselected_with_keywords[df_unselected_with_keywords.topic_probs>0.01]
-
-# %%
-
-# %% [markdown]
-# ## Hydrogen: false positive and negative check
-
-# %%
-category = "Hydrogen & Fuel Cells"
-categories_to_narrow_topics[category]
-
-# %%
-# Narrow topic probability
-doc_df_ = doc_df[["doc_id"]].copy()
-doc_df_["topic_probs_1"] = narrow_topic_probs[:, 94]
-doc_df_["topic_probs_2"] = narrow_topic_probs[:, 130]
-# All documents with the keywords
-# df_with_keywords = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords_old = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords = get_docs_with_keyphrases_2(narrow_ref_keywords[category])
-print(len(df_with_keywords_old), len(df_with_keywords))
-
-# Above-threshold, selected documents that have the keyword
-df_selected = selected_docs_df_all[selected_docs_df_all.category == category]
-df_selected["has_keyword"] = 0
-df_selected.loc[
-    df_selected.doc_id.isin(df_with_keywords.doc_id.to_list()), "has_keyword"
-] = 1
-# Documents with keywords that were not selected
-df_unselected_with_keywords = df_with_keywords[
-    df_with_keywords.doc_id.isin(df_selected.doc_id.to_list()) == False
-]
-df_unselected_with_keywords = df_unselected_with_keywords.merge(doc_df_, how="left")
-
-# %%
-df_ = df_selected[["doc_id", "title", "description", "source"]]
-df_ = df_.merge(doc_df_)
-df_["manual_ok"] = 0
-df_new = df_unselected_with_keywords
-df_new["manual_ok"] = 0
-df_ = df_.append(df_new, ignore_index=True)
-df_ = df_.merge(clustering[["doc_id", "cluster_keywords"]])
-df_["category"] = category
-
-data_for_labelling = []
-for i, row in df_.iterrows():
-    s = f"CATEGORY: {category}\n\n"
-    s += f"doc_id: {row.doc_id}\n\nprobability: {np.round(row.topic_probs_1,4)}\n\n{row.title}\n------\n{row.description}\n------"
-    s += f"\n\ntop2vec keywords: {row.cluster_keywords}"
-    data_for_labelling.append(s)
-df_["text"] = data_for_labelling
-
-# %%
-df_new_cb = df_[(df_.source == "cb") & (df_.manual_ok == 0)]
-df_new_cb.to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Aug23_cb_{category}.csv",
-    index=False,
-)
-
-df_new_cb[df_new_cb.doc_id.isin(manual_cb_review.doc_id.to_list()) == False].to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Sep1_cb_{category}.csv",
-    index=False,
-)
-
-DF_NEW[category] = df_[
-    (df_.source == "gtr") | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-
-
-# %%
-print(len(DF_NEW[category]))
-DF_NEW[category] = df_[
-    (
-        (df_.source == "gtr")
-        & ((df_.manual_ok == 1) | df_.doc_id.isin(df_with_keywords.doc_id.to_list()))
-    )
-    | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-len(DF_NEW[category])
-
-# %%
-len(df_selected), df_selected.has_keyword.sum(), len(df_unselected_with_keywords)
-
-# %%
-# len(df_unselected_with_keywords[
-#     (df_unselected_with_keywords.topic_probs_1>0.05) |
-#     (df_unselected_with_keywords.topic_probs_2>0.05)
-# ])
-
-# %%
-# DF_NEW.keys()
-
-# %%
-# iss_io.save_pickle(DF_NEW, PROJECT_DIR / 'outputs/data/results_august/reference')
-
-# %%
-# DF_NEW
-
-# %%
-cats
-
-# %% [markdown]
-# ## Bioenergy: false positive and negative check
-
-# %%
-category = "Bioenergy"
-categories_to_narrow_topics[category]
-
-# %%
-category_reviewed = df_manual_labels[df_manual_labels.category == category]
-category_reviewed_hits = category_reviewed[category_reviewed.sentiment == "Keep"]
-category_reviewed_misses = category_reviewed[category_reviewed.sentiment == "Discard"]
-
-# %%
-# Narrow topic probability
-doc_df_ = doc_df[["doc_id"]].copy()
-doc_df_["topic_probs_1"] = narrow_topic_probs[:, 68]
-doc_df_["topic_probs_2"] = narrow_topic_probs[:, 92]
-doc_df_["topic_probs_3"] = narrow_topic_probs[:, 93]
-# All documents with the keywords
-# df_with_keywords = get_docs_with_keyphrases(narrow_ref_keywords[category])
-
-df_with_keywords_old = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords = get_docs_with_keyphrases_2(narrow_ref_keywords[category])
-print(len(df_with_keywords_old), len(df_with_keywords))
-
-# Above-threshold, selected documents that have the keyword
-df_selected = selected_docs_df_all[
-    selected_docs_df_all.doc_id.isin(category_reviewed.doc_id.to_list())
-]
-df_selected["has_keyword"] = 0
-df_selected.loc[
-    df_selected.doc_id.isin(df_with_keywords.doc_id.to_list()), "has_keyword"
-] = 1
-# Documents with keywords that were not selected
-df_unselected_with_keywords = df_with_keywords[
-    df_with_keywords.doc_id.isin(df_selected.doc_id.to_list()) == False
-]
-df_unselected_with_keywords = df_unselected_with_keywords.merge(doc_df_, how="left")
-
-# %%
-df_ = category_reviewed_hits[["doc_id", "title", "description", "source"]]
-df_ = df_.merge(doc_df_)
-df_["manual_ok"] = 1
-df_new = df_unselected_with_keywords
-df_new["manual_ok"] = 0
-df_ = df_.append(df_new, ignore_index=True)
-df_ = df_.merge(clustering[["doc_id", "cluster_keywords"]])
-df_["category"] = category
-
-data_for_labelling = []
-for i, row in df_.iterrows():
-    s = f"CATEGORY: {category}\n\n"
-    s += f"doc_id: {row.doc_id}\n\nprobability: {np.round(row.topic_probs_1,4)}\n\n{row.title}\n------\n{row.description}\n------"
-    s += f"\n\ntop2vec keywords: {row.cluster_keywords}"
-    data_for_labelling.append(s)
-df_["text"] = data_for_labelling
-
-# %%
-df_new_cb = df_[(df_.source == "cb") & (df_.manual_ok == 0)]
-df_new_cb.to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Aug23_cb_{category}.csv",
-    index=False,
-)
-
-
-DF_NEW[category] = df_[
-    (df_.source == "gtr") | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-
-
-# %%
-print(len(DF_NEW[category]))
-DF_NEW[category] = df_[
-    (
-        (df_.source == "gtr")
-        & ((df_.manual_ok == 1) | df_.doc_id.isin(df_with_keywords.doc_id.to_list()))
-    )
-    | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-len(DF_NEW[category])
-
-# %% [markdown]
-# ### Carbon capture
-
-# %%
-category = "Carbon Capture & Storage"
-categories_to_narrow_topics[category]
-
-# %%
-# Narrow topic probability
-doc_df_ = doc_df[["doc_id"]].copy()
-doc_df_["topic_probs"] = narrow_topic_probs[:, 39]
-# All documents with the keywords
-# df_with_keywords = get_docs_with_keyphrases(narrow_ref_keywords[category])
-
-df_with_keywords_old = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords = get_docs_with_keyphrases_2(narrow_ref_keywords[category])
-print(len(df_with_keywords_old), len(df_with_keywords))
-# Above-threshold, selected documents that have the keyword
-df_selected = selected_docs_df_all[selected_docs_df_all.category == category]
-df_selected["has_keyword"] = 0
-df_selected.loc[
-    df_selected.doc_id.isin(df_with_keywords.doc_id.to_list()), "has_keyword"
-] = 1
-# Documents with keywords that were not selected
-df_unselected_with_keywords = df_with_keywords[
-    df_with_keywords.doc_id.isin(df_selected.doc_id.to_list()) == False
-]
-df_unselected_with_keywords = df_unselected_with_keywords.merge(doc_df_, how="left")
-
-# %%
-df_ = df_selected[["doc_id", "title", "description", "source"]]
-df_ = df_.merge(doc_df_)
-df_["manual_ok"] = 0
-df_new = df_unselected_with_keywords
-df_new["manual_ok"] = 0
-df_ = df_.append(df_new, ignore_index=True)
-df_ = df_.merge(clustering[["doc_id", "cluster_keywords"]])
-df_["category"] = category
-
-data_for_labelling = []
-for i, row in df_.iterrows():
-    s = f"CATEGORY: {category}\n\n"
-    s += f"doc_id: {row.doc_id}\n\nprobability: {np.round(row.topic_probs,4)}\n\n{row.title}\n------\n{row.description}\n------"
-    s += f"\n\ntop2vec keywords: {row.cluster_keywords}"
-    data_for_labelling.append(s)
-df_["text"] = data_for_labelling
-
-# %%
-df_new_cb = df_[(df_.source == "cb") & (df_.manual_ok == 0)]
-df_new_cb.to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Aug23_cb_{category}.csv",
-    index=False,
-)
-
-df_new_cb[df_new_cb.doc_id.isin(manual_cb_review.doc_id.to_list()) == False].to_csv(
-    PROJECT_DIR / f"outputs/data/results_august/aux/review/Sep1_cb_{category}.csv",
-    index=False,
-)
-
-DF_NEW[category] = df_[
-    (df_.source == "gtr") | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-
-
-# %%
-print(len(DF_NEW[category]))
-DF_NEW[category] = df_[
-    (
-        (df_.source == "gtr")
-        & ((df_.manual_ok == 1) | df_.doc_id.isin(df_with_keywords.doc_id.to_list()))
-    )
-    | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-len(DF_NEW[category])
-
-# %% [markdown]
-# ### Heating
-
-# %%
-CB_DOCS_ALL = pd.read_csv(
-    PROJECT_DIR / "outputs/data/results_august/checked_heating_tech_CB.csv"
-)
-GTR_DOCS_ALL = pd.read_csv(
-    PROJECT_DIR / "outputs/data/results_august/checked_heating_tech_GTR.csv"
-)
-
-# %%
-# CB_DOCS_ALL[CB_DOCS_ALL.title.str.contains('Muovitech')]
-
-# %%
-doc_id_checked = GTR_DOCS_ALL.doc_id.to_list() + CB_DOCS_ALL.doc_id.to_list()
-
-# %%
-len(doc_id_checked)
-
-# %%
-category = "Heating (all other)"
-
-# %%
-# Narrow topic probability
-doc_df_ = doc_df[["doc_id"]].copy()
-doc_df_["topic_probs"] = narrow_topic_probs[:, 19]
-# All documents with the keywords
-# df_with_keywords = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords_old = get_docs_with_keyphrases(narrow_ref_keywords[category])
-df_with_keywords = get_docs_with_keyphrases_2(narrow_ref_keywords[category])
-print(len(df_with_keywords_old), len(df_with_keywords))
-# Above-threshold, selected documents that have the keyword
-df_selected = selected_docs_df_all[selected_docs_df_all.doc_id.isin(doc_id_checked)]
-df_selected["has_keyword"] = 0
-df_selected.loc[
-    df_selected.doc_id.isin(df_with_keywords.doc_id.to_list()), "has_keyword"
-] = 1
-# Documents with keywords that were not selected
-df_unselected_with_keywords = df_with_keywords[
-    df_with_keywords.doc_id.isin(df_selected.doc_id.to_list()) == False
-]
-df_unselected_with_keywords = df_unselected_with_keywords.merge(doc_df_, how="left")
-
-# %%
-df_ = df_selected[["doc_id", "title", "description", "source"]]
-df_ = df_.merge(doc_df_)
-df_["manual_ok"] = 1
-df_new = df_unselected_with_keywords
-df_new["manual_ok"] = 0
-df_ = df_.append(df_new, ignore_index=True)
-df_ = df_.merge(clustering[["doc_id", "cluster_keywords"]])
-df_["category"] = category
-
-data_for_labelling = []
-for i, row in df_.iterrows():
-    s = f"CATEGORY: {category}\n\n"
-    s += f"doc_id: {row.doc_id}\n\nprobability: {np.round(row.topic_probs,4)}\n\n{row.title}\n------\n{row.description}\n------"
-    s += f"\n\ntop2vec keywords: {row.cluster_keywords}"
-    data_for_labelling.append(s)
-df_["text"] = data_for_labelling
-
-# %%
-# df_new_cb = df_[(df_.source=='cb') & (df_.manual_ok==0)]
-# df_new_cb.to_csv(PROJECT_DIR / f'outputs/data/results_august/aux/review/Aug23_cb_{category}.csv', index=False)
-DF_NEW[category] = df_[
-    (df_.source == "gtr") | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-
-
-# %%
-print(len(DF_NEW[category]))
-DF_NEW[category] = df_[
-    (
-        (df_.source == "gtr")
-        & ((df_.manual_ok == 1) | df_.doc_id.isin(df_with_keywords.doc_id.to_list()))
-    )
-    | ((df_.source == "cb") & (df_.manual_ok == 1))
-]
-len(DF_NEW[category])
-
-# %% [markdown]
-# ## Combine and save
-
-# %%
-DF_REF = DF_NEW.copy()
-
-# %%
-DF_REF[cat].head(1)
-
-# %%
-manual_cb_review = pd.read_csv(
-    PROJECT_DIR
-    / "outputs/data/results_august/aux/project-11-at-2021-08-24-17-28-6e9a63e0.csv"
-)
-manual_cb_review = pd.read_csv(
-    PROJECT_DIR
-    / "outputs/data/results_august/aux/project-11-at-2021-09-01-10-01-24e5dd2e.csv"
-)
-
-# %%
-cat = "Wind & Offshore"
-df = manual_cb_review[
-    (manual_cb_review.category == cat) & (manual_cb_review.sentiment == "Keep")
-]
-df["manual_ok"] = 1
-df = df[
-    [
-        "doc_id",
-        "title",
-        "description",
-        "source",
-        "topic_probs",
-        "manual_ok",
-        "cluster_keywords",
-        "text",
-    ]
-]
-DF_REF[cat] = DF_REF[cat].append(df, ignore_index=True)
-
-# %%
-DF_REF[cat].source.unique()
-
-# %%
-cat = "Solar"
-df = manual_cb_review[
-    (manual_cb_review.category == cat) & (manual_cb_review.sentiment == "Keep")
-]
-df["manual_ok"] = 1
-df = df[
-    [
-        "doc_id",
-        "title",
-        "description",
-        "source",
-        "topic_probs",
-        "manual_ok",
-        "cluster_keywords",
-        "text",
-    ]
-]
-DF_REF[cat] = DF_REF[cat].append(df, ignore_index=True)
-
-# %%
-cat = "Batteries"
-df = manual_cb_review[
-    (manual_cb_review.category == cat) & (manual_cb_review.sentiment == "Keep")
-]
-df["manual_ok"] = 1
-df = df[
-    [
-        "doc_id",
-        "title",
-        "description",
-        "source",
-        "topic_probs",
-        "manual_ok",
-        "cluster_keywords",
-        "text",
-    ]
-]
-DF_REF[cat] = DF_REF[cat].append(df, ignore_index=True)
-
-# %%
-DF_REF["Batteries"].head(1)
-
-# %%
-cat = "Hydrogen & Fuel Cells"
-df = manual_cb_review[
-    (manual_cb_review.category == cat) & (manual_cb_review.sentiment == "Keep")
-]
-df["manual_ok"] = 1
-df = df[
-    [
-        "doc_id",
-        "title",
-        "description",
-        "source",
-        "topic_probs_1",
-        "topic_probs_2",
-        "manual_ok",
-        "cluster_keywords",
-        "text",
-    ]
-]
-DF_REF[cat] = DF_REF[cat].append(df, ignore_index=True)
-
-
-# %%
-DF_REF[cat]["topic_probs"] = 0
-DF_REF[cat] = DF_REF[cat].reset_index(drop=True)
-for i, row in DF_REF[cat].iterrows():
-    DF_REF[cat].loc[i, "topic_probs"] = max(row["topic_probs_1"], row["topic_probs_2"])
-
-# %%
-DF_REF[cat] = DF_REF[cat].drop(["topic_probs_1", "topic_probs_2"], axis=1)
-
-# %%
-DF_REF[cat] = DF_REF[cat][
-    [
-        "doc_id",
-        "title",
-        "description",
-        "source",
-        "topic_probs",
-        "manual_ok",
-        "cluster_keywords",
-        "text",
-    ]
-]
-
-# %%
-DF_REF[cat].head(1)
-
-# %%
-cat = "Bioenergy"
-df = manual_cb_review[
-    (manual_cb_review.category == cat) & (manual_cb_review.sentiment == "Keep")
-]
-df["manual_ok"] = 1
-df = df[
-    [
-        "doc_id",
-        "title",
-        "description",
-        "source",
-        "topic_probs_1",
-        "topic_probs_2",
-        "topic_probs_3",
-        "manual_ok",
-        "cluster_keywords",
-        "text",
-    ]
-]
-
-DF_REF[cat]["topic_probs"] = 0
-DF_REF[cat] = DF_REF[cat].reset_index(drop=True)
-for i, row in DF_REF[cat].iterrows():
-    DF_REF[cat].loc[i, "topic_probs"] = max(
-        row["topic_probs_1"], row["topic_probs_2"], row["topic_probs_3"]
-    )
-DF_REF[cat] = DF_REF[cat].drop(
-    ["topic_probs_1", "topic_probs_2", "topic_probs_3"], axis=1
-)
-DF_REF[cat] = DF_REF[cat][
-    [
-        "doc_id",
-        "title",
-        "description",
-        "source",
-        "topic_probs",
-        "manual_ok",
-        "cluster_keywords",
-        "text",
-    ]
-]
-
-DF_REF[cat] = DF_REF[cat].append(df, ignore_index=True)
-
-# %%
-cat = "Carbon Capture & Storage"
-df = manual_cb_review[
-    (manual_cb_review.category == cat) & (manual_cb_review.sentiment == "Keep")
-]
-df["manual_ok"] = 1
-df = df[
-    [
-        "doc_id",
-        "title",
-        "description",
-        "source",
-        "topic_probs",
-        "manual_ok",
-        "cluster_keywords",
-        "text",
-    ]
-]
-DF_REF[cat] = DF_REF[cat].append(df, ignore_index=True)
-
-# %%
-iss_io.save_pickle(
-    DF_REF, PROJECT_DIR / "outputs/data/results_august/reference_category_data_2.p"
-)
-
-# %% [markdown]
-# # Fuel poverty & social
-
-# %%
-GTR_DOCS = pd.read_csv(PROJECT_DIR / "outputs/data/results_august/FINAL_TABLES_GTR.csv")
-CB_DOCS = pd.read_csv(PROJECT_DIR / "outputs/data/results_august/FINAL_TABLES_CB.csv")
-
-# %% [markdown]
-# ## Keywords
-
-# %%
-keywords_fuel = [["fuel poverty"], ["fuel poor"]]
-
-keywords_energy = [["energy poverty"], ["energy poor"]]
-
-keywords_social = [["social hous"], ["social landlord"]]
-
-# %%
-df_with_keywords = get_docs_with_keyphrases_2(keywords_fuel)
-len(df_with_keywords)
-
-# %%
-# df_with_keywords[df_with_keywords.doc_id.isin(df_with_keywords_.doc_id.to_list())]
-
-# %%
-df_with_keywords = get_docs_with_keyphrases_2(keywords_energy)
-len(df_with_keywords)
-
-# %%
-# df_with_keywords.to_csv(PROJECT_DIR / 'outputs/data/results_august/fuel_poverty_docs.csv', index=False)
-
-# %%
-df_fp = pd.read_csv(
-    PROJECT_DIR / "outputs/data/results_august/fuel_poverty_docs_checked.csv"
-)
-df_fp = df_fp[df_fp.hit != 0]
-
-# %%
-df_lch_eem_gtr = GTR_DOCS[GTR_DOCS.tech_category.isin(["LCH & EEM"])]
-fp_gtr = df_lch_eem_gtr[df_lch_eem_gtr.doc_id.isin(df_fp.doc_id.to_list())]
-
-# %%
-len(fp_gtr) / len(df_lch_eem_gtr)
-
-# %%
-GTR_DOCS[GTR_DOCS.doc_id.isin(df_fp.doc_id.to_list())].groupby(
-    "tech_category"
-).count().sort_values("doc_id")
-
-# %%
-7 / len(fp_gtr)
-
-# %%
-5 / len(fp_gtr)
-
-# %%
-df_lch_eem_cb = CB_DOCS[CB_DOCS.tech_category.isin(["LCH & EEM"])]
-fp_cb = df_lch_eem_cb[df_lch_eem_cb.doc_id.isin(df_fp.doc_id.to_list())]
-
-# %%
-len(fp_cb) / len(df_lch_eem_cb)
-
-# %%
-fp_cb
-
-# %%
-df_with_keywords
-
-# %%
-## Topics
-
-# %% [markdown]
-# # Enabling technologies
-
-# %%
-doc_topic_dists = get_doc_topic_dists(mdl)
-doc_topic_dists.shape
-
-# %%
-GTR_DOCS_ALL = pd.read_csv(
-    PROJECT_DIR / "outputs/data/results_august/checked_heating_tech.csv"
-)
-
-# %%
-# df = GTR_DOCS_ALL[GTR_DOCS_ALL.tech_category=='Building insulation']
-df = GTR_DOCS_ALL
-doc_df_index = doc_df[["doc_id"]].reset_index().rename(columns={"index": "i"})
-df = df.merge(doc_df_index, how="left")
-df.head(3)
-
-# %%
-doc_topic_dists.shape
-
-# %%
-df["year"] = df["start"].apply(iss.convert_date_to_year)
-
-# %%
-
-# %%
-doc_df_prob = doc_df[["doc_id"]]
-doc_df_prob["prob"] = np.max(
-    doc_topic_dists[:, [10, 17, 18, 85, 89, 129, 145]], axis=1
-)  # doc_topic_dists[:, 17]
-df_t = df.merge(doc_df_prob).sort_values("prob")
-
-alt.Chart(df_t[df_t.tech_category == "Geothermal energy"]).mark_circle(size=18).encode(
-    x="year:O",
-    y=alt.X("prob", scale=alt.Scale(type="log")),
-    color="tech_category",
-    tooltip=["tech_category", "title", "description"],
-).interactive()
-
-# %%
-plt.scatter(df_t.year, df_t.prob)
-plt.y_scale("log")
-
-# %%
-# sorted(df.year.unique())
-
-# %%
-year_dists = []
-years = sorted(df.year.unique())
-for year in years:
-    df_y = df[df.year == year]
-    doc_topic_dists_y = doc_topic_dists[df_y.i.to_list(), :]
-    doc_topic_dists_y_t = doc_topic_dists_y[:, 18]
-    year_dists.append(doc_topic_dists_y_t)
-
-# %%
-top_topic_pctile = [np.percentile(y, 50) for y in year_dists]
-
-# %%
-sns.lineplot(x=years, y=top_topic_pctile)
-
-# %%
-# Check topic intensity
-
-# %%
-# doc_topic_dists[0,:]
-
-# %% [markdown]
-# # Estimate accuracy
-
-# %%
-df_manual_labels = pd.read_csv(
-    "/Users/karliskanders/Downloads/project-7-at-2021-08-20-11-17-039f3681.csv"
-)
-
-# %%
-category = "Solar"
-categories_to_narrow_topics[category]
-
-# %%
-# Narrow topic probability
-doc_df_ = doc_df[["doc_id"]].copy()
-doc_df_["topic_probs"] = narrow_topic_probs[:, 2]
-# All documents with the keywords
-df_with_keywords = get_docs_with_keyphrases(narrow_ref_keywords[category])
-# Above-threshold, selected documents that have the keyword
-df_selected = selected_docs_df_all[selected_docs_df_all.category == category]
-df_selected["has_keyword"] = 0
-df_selected.loc[
-    df_selected.doc_id.isin(df_with_keywords.doc_id.to_list()), "has_keyword"
-] = 1
-# Documents with keywords that were not selected
-df_unselected_with_keywords = df_with_keywords[
-    df_with_keywords.doc_id.isin(df_selected.doc_id.to_list()) == False
-]
-df_unselected_with_keywords = df_unselected_with_keywords.merge(doc_df_, how="left")
-
-# %%
-df = df_manual_labels[
-    (df_manual_labels.category == category) & (df_manual_labels.source == "gtr")
-].merge(doc_df_)
-sns.displot(df[df.sentiment == "Keep"].topic_probs.to_list())
-sns.displot(df[df.sentiment == "Discard"].topic_probs.to_list())
-plt.show()
-
-# %%
-# df[df.sentiment=='Keep'].sort_values('topic_probs').head(15)
 
 # %%
