--- innovation_sweet_spots/analysis/prototyping/23_verified_trends_v3.py
@@ -1,27 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     cell_metadata_filter: -all
-#     comment_magics: true
-#     text_representation:
-#       extension: .py
-#       format_name: percent
-#       format_version: '1.3'
-#       jupytext_version: 1.11.3
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-# %%
-import innovation_sweet_spots.analysis.analysis_utils as iss
-from innovation_sweet_spots import PROJECT_DIR, logging, config
-from innovation_sweet_spots.getters import gtr, crunchbase, guardian
-import innovation_sweet_spots.analysis.topic_analysis as iss_topics
-import pandas as pd
-import numpy as np
-
-# %%
-
-# %%
